forced prettier:fix run

+++ contracts/interfaces/IVersionRegistry.sol
@@ -2,9 +2,13 @@
 pragma solidity ^0.6.12;
 
 interface IVersionRegistry {
-
     //event emitted whenever a version is added
-    event VersionAdded(bytes32 indexed id, bytes32 version, string value, uint time);
+    event VersionAdded(
+        bytes32 indexed id,
+        bytes32 version,
+        string value,
+        uint256 time
+    );
 
     //event emitted whenever a version is canceled
     event VersionCanceled(bytes32 indexed id, bytes32 version, string reason);
@@ -15,10 +19,18 @@ interface IVersionRegistry {
      * @param version the new version to add (32-byte string)
      * @param value value to attach to this version
      */
-    function addVersion(bytes32 id, bytes32 version, string calldata value) external;
+    function addVersion(
+        bytes32 id,
+        bytes32 version,
+        string calldata value
+    ) external;
 
     /**
      * cancel a version.
      */
-    function cancelVersion(bytes32 id, bytes32 version, string calldata reason) external;
+    function cancelVersion(
+        bytes32 id,
+        bytes32 version,
+        string calldata reason
+    ) external;
 }
